% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinoplot.R
\name{dinoPlot}
\alias{dinoPlot}
\title{Plot a dinosaur}
\usage{
dinoPlot(col = "black", add = T)
}
\arguments{
\item{col}{colour of the dinosaur}

\item{add}{Add to existing plot (T/F)}
}
\value{
Dino plot
}
\description{
This function takes the data from datasauRus and
plot the dinosaur with you favorite color
}
